refactor: separates linux and macos implementations.
Files:
- restart_handler_unix.cpp/hpp - functions common to linux and macOS.
- restart_handler_linux.cpp - linux implementation
- restart_handler_darwin.cpp - macOS implementation
