+++ src/agent/restart_handler/src/restart_handler_darwin.cpp
@@ -1,0 +1,14 @@
+#include <restart_handler_unix.hpp>
+
+namespace restart_handler
+{
+    bool UsingSystemctl()
+    {
+        // TODO: Implement this function for macOS
+    }
+
+    boost::asio::awaitable<module_command::CommandExecutionResult> RestartWithSystemd()
+    {
+        // TODO: Implement this function for macOS
+    }
+} // namespace restart_handler
